feat: Convert check-for-updates command to an auto-update command with more options
In order to enable auto updates to be easily enabled/disabled, I've converted
the `check-for-updates` command to an `auto-update` command which can be used to
enable and disable automatic updates in addition to manually triggering an update check.

Auto-updates are enabled/disabled by creating an empty `.disable-auto-updates` file in
the root of the repository. That file is checked before an auto-update is performed, and
if it exists no update check happens.

Updates can still be manually performed with `e auto-update check` or just `e auto-update`.

The old `e check-for-updates` command is kept as an alias for backward compatibility.
